+++ django_deovi/models/directory.py
@@ -12,13 +12,15 @@ from smart_media.modelfields import SmartMediaField
 from smart_media.mixins import SmartFormatMixin
 from smart_media.signals import auto_purge_files_on_change, auto_purge_files_on_delete
 
+from ..utils.text import normalize_text
+
 
 class Directory(SmartFormatMixin, models.Model):
     """
     A directory container to hold MediaFile objects.
 
     TODO:
-    * 'genres' from payload should be a many2many, we ignore this feature for now;
+    * New field 'genres' collected from payload, it should be a many2many;
     * Stored payload should not include the mediafile list (loader should pop it away);
     * Payload should contains something like: ::
 
@@ -171,6 +173,9 @@ class Directory(SmartFormatMixin, models.Model):
 
     def __str__(self):
         return self.title or self.directory_name()
+
+    def normalized_title(self):
+        return normalize_text(self.title)
 
     def get_absolute_url(self):
         """