+++ sandbox/settings/demo.py
@@ -7,16 +7,19 @@ from sandbox.settings.base import *  # noqa: F403
 
 DEBUG = True
 
-TEMPLATES[0]['OPTIONS']['debug'] = DEBUG  # noqa: F405
+TEMPLATES[0]["OPTIONS"]["debug"] = DEBUG  # noqa: F405
 
 DATABASES = {
     "default": {
         "ENGINE": "django.db.backends.sqlite3",
-        "NAME": join(VAR_PATH, "db", "db.sqlite3"),  # noqa: F405
+        "NAME": VAR_PATH / "db" / "db.sqlite3",  # noqa: F405
     }
 }
 
+LANGUAGE_CODE = "fr"
 TIME_ZONE = "Europe/Paris"
+
+DEOVI_INDEXES_DEBUG = True
 
 # Import local settings if any
 try: