+++ sandbox/settings/tests.py
@@ -13,4 +13,4 @@ DATABASES = {
 
 # Media directory dedicated to tests to avoid polluting other environment
 # media directory
-MEDIA_ROOT = join(VAR_PATH, "media-tests")  # noqa: F405
+MEDIA_ROOT = VAR_PATH / "media-tests"  # noqa: F405